Create java problem 1
This example creates a program to let a first grader practice additions. The program randomly generates two single-digit integers number1 and number2 and displays a question such as “What is 9-+7?” to the student. After the student types the answer, the program displays a message to indicate whether the answer is true or false.