Problem 92
Design a class named Student in Java that contains:
A public int data field named id for the student.
A public double data field named cgpa for the student.
A public string data field named name for the student.
A constructor that creates a student with the specified id, cgpa and name.
A method named viewInfo() that prints all data related to a student.
Write a test program that creates n (an integer variable) student objects. Take input in the data fields for each object. If the cgpa is more than 4.00 or less than 0.00 then throw a custom InvalidCGPAException. Sort the student objects in descending order based on the character at the index 1 of the name data field and show the sorted student records. Mention all the other exceptions in details.