Problem 102
Write a Java program to create a directory (named University) in E drive of your computer and a .txt file (named Student) inside that directory. Enter n( an integer variable, for which you have to
take input from console) lines in the file. Each line in the file consists of a student’s first name, last name, year of admission, and id. The student’s first name and last name for the ith line are
FirstNamei and LastNamei. The year of admission is a randomly generated number between 2012-2024. The id is randomly generated as a 8 digit number, where the first 4 digits are the
year of admission and the last 4 are random and non repetitive between other students. Save the file Student.txt and read data from the into the console. Here are some sample data:
FirstName1 LastName1 2013 20131234
FirstName2 LastName2 2019 20192345
……….
………..
………
……..
……..
FirstName1000 LastName1000 2020 20206789